add dao for goods

+++ webshop/src/main/java/ua/nung/edu/pz/dao/repository/UserRepository.java
@@ -18,7 +18,7 @@ public class UserRepository {
 				Statement statement = connection.createStatement();
 				ResultSet resultSet = statement.executeQuery(sql);) {
 			if (resultSet.next()) {
-				user.setId(resultSet.getInt("id"));
+				user.setId(resultSet.getLong("id"));
 				user.setEmail(resultSet.getString("email"));
 				user.setPassword(resultSet.getString("password"));
 				user.setDisplayName(resultSet.getString("displayName"));
